TAJO-838_3

+++ tajo-common/src/main/java/org/apache/tajo/OverridableConf.java
@@ -171,7 +171,7 @@ public class OverridableConf extends KeyValueSet {
   }
 
   public float getFloat(ConfigKey key) {
-    return getLong(key, null);
+    return getFloat(key, null);
   }
 
   public void put(ConfigKey key, String val) {
